chore(mobile): move graph stuff related to only mobile to suite-native
