+++ suite-common/graph/src/hooks.ts
@@ -1,28 +1,13 @@
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useSelector } from 'react-redux';
 
 import { roundToNearestMinutes, subHours } from 'date-fns';
 
 import { FiatCurrencyCode } from '@suite-common/suite-config';
-import { selectAccountByKey, selectAccounts } from '@suite-common/wallet-core';
-import { AccountKey } from '@suite-common/wallet-types';
 
 import { getMultipleAccountBalanceHistoryWithFiat } from './graphDataFetching';
 import { AccountItem, FiatGraphPoint, FiatGraphPointWithCryptoBalance } from './types';
 
-const getTimeFrameForHistoryHours = (hoursToHistory: number | null) => {
-    const endOfTimeFrameDate = roundToNearestMinutes(new Date(), {
-        nearestTo: 10,
-        roundingMethod: 'floor',
-    });
-    const startOfTimeFrameDate = hoursToHistory
-        ? subHours(endOfTimeFrameDate, hoursToHistory)
-        : null;
-
-    return { startOfTimeFrameDate, endOfTimeFrameDate };
-};
-
-type CommonUseGraphParams = {
+export type CommonUseGraphParams = {
     fiatCurrency: FiatCurrencyCode;
 };
 
@@ -88,64 +73,15 @@ export const useGraphForAccounts = ({
     return { graphPoints, isLoading, error, refetch };
 };
 
-export const useGraphForSingleAccount = ({
-    accountKey,
-    fiatCurrency,
-}: CommonUseGraphParams & { accountKey: AccountKey }) => {
-    const account = useSelector((state: any) => selectAccountByKey(state, accountKey));
-    const [hoursToHistory, setHoursToHistory] = useState<number | null>(24);
-    const { startOfTimeFrameDate, endOfTimeFrameDate } = useMemo(
-        () => getTimeFrameForHistoryHours(hoursToHistory),
-        [hoursToHistory],
-    );
+export const useGetTimeFrameForHistoryHours = (hoursToHistory: number | null) =>
+    useMemo(() => {
+        const endOfTimeFrameDate = roundToNearestMinutes(new Date(), {
+            nearestTo: 10,
+            roundingMethod: 'floor',
+        });
+        const startOfTimeFrameDate = hoursToHistory
+            ? subHours(endOfTimeFrameDate, hoursToHistory)
+            : null;
 
-    const accounts = useMemo(() => {
-        if (!account) return [];
-        return [
-            {
-                coin: account.symbol,
-                descriptor: account.descriptor,
-            },
-        ] as AccountItem[];
-    }, [account]);
-
-    return {
-        ...useGraphForAccounts({
-            accounts,
-            fiatCurrency,
-            startOfTimeFrameDate,
-            endOfTimeFrameDate,
-        }),
-        hoursToHistory,
-        setHoursToHistory,
-    };
-};
-
-export const useGraphForAllAccounts = ({ fiatCurrency }: CommonUseGraphParams) => {
-    const accounts = useSelector(selectAccounts);
-    const [hoursToHistory, setHoursToHistory] = useState<number | null>(24);
-    const { startOfTimeFrameDate, endOfTimeFrameDate } = useMemo(
-        () => getTimeFrameForHistoryHours(hoursToHistory),
-        [hoursToHistory],
-    );
-
-    const accountItems = useMemo(
-        () =>
-            accounts.map(account => ({
-                coin: account.symbol,
-                descriptor: account.descriptor,
-            })),
-        [accounts],
-    );
-
-    return {
-        ...useGraphForAccounts({
-            accounts: accountItems,
-            fiatCurrency,
-            startOfTimeFrameDate,
-            endOfTimeFrameDate,
-        }),
-        hoursToHistory,
-        setHoursToHistory,
-    };
-};
+        return { endOfTimeFrameDate, startOfTimeFrameDate };
+    }, [hoursToHistory]);
