+++ suite-native/graph/package.json
@@ -15,6 +15,7 @@
         "@shopify/react-native-skia": "0.1.162",
         "@suite-common/formatters": "workspace:*",
         "@suite-common/graph": "workspace:*",
+        "@suite-common/wallet-core": "workspace:*",
         "@suite-common/wallet-types": "workspace:*",
         "@suite-native/atoms": "workspace:*",
         "@suite-native/test-utils": "workspace:*",
