+++ suite-native/graph/src/index.ts
@@ -1,3 +1,4 @@
 export * from './components/Graph';
 export * from './components/TimeSwitch';
 export * from './utils';
+export * from './hooks';
