+++ suite-native/graph/tsconfig.json
@@ -7,6 +7,9 @@
         },
         { "path": "../../suite-common/graph" },
         {
+            "path": "../../suite-common/wallet-core"
+        },
+        {
             "path": "../../suite-common/wallet-types"
         },
         { "path": "../atoms" },
