+++ suite-native/module-accounts/tsconfig.json
@@ -5,7 +5,6 @@
         {
             "path": "../../suite-common/formatters"
         },
-        { "path": "../../suite-common/graph" },
         {
             "path": "../../suite-common/suite-types"
         },
