+++ suite-native/module-home/src/components/PortfolioGraph.tsx
@@ -3,8 +3,7 @@ import { useSelector } from 'react-redux';
 
 import { useAtom } from 'jotai';
 
-import { useGraphForAllAccounts } from '@suite-common/graph';
-import { enhanceGraphPoints, Graph, TimeSwitch } from '@suite-native/graph';
+import { useGraphForAllAccounts, enhanceGraphPoints, Graph, TimeSwitch } from '@suite-native/graph';
 import { selectFiatCurrency } from '@suite-native/module-settings';
 
 import {
