+++ suite-native/module-home/tsconfig.json
@@ -8,7 +8,6 @@
         {
             "path": "../../suite-common/formatters"
         },
-        { "path": "../../suite-common/graph" },
         {
             "path": "../../suite-common/suite-utils"
         },
